requirements.txt
streamlit==1.32.2
openai==1.12.0
python-dotenv==1.0.1    